Fixed more compiler warnings.
git-svn-id: https://gforge.sci.utah.edu/svn/nektar/trunk@3457 305cdda6-5ce1-45b3-a98d-dfc68c8b3305

+++ library/LibUtilities/BasicUtils/SessionReader.cpp
@@ -348,7 +348,7 @@ namespace Nektar
             ParameterMap::const_iterator paramMapIter = m_parameters.find(vName);
             ASSERTL0(paramMapIter != m_parameters.end(),
                     "Required parameter '" + pName + "' not specified in session.");
-            pVar = floor(paramMapIter->second);
+            pVar = (int)floor(paramMapIter->second);
         }
 
 
@@ -361,7 +361,7 @@ namespace Nektar
             ParameterMap::const_iterator paramMapIter = m_parameters.find(vName);
             if(paramMapIter != m_parameters.end())
             {
-                pVar = floor(paramMapIter->second);
+                pVar = (int)floor(paramMapIter->second);
             }
             else
             {
